Move CohortsAlumniPage from component to containers and create a new folder for Specific Alumnus Page with add it's route into ROUTES
