+++ client/src/constants/routes.js
@@ -18,6 +18,7 @@ export const ALUMNI_PAGE = '/alumni';
 export const SPECIFIC_COHORT_PAGE = '/cohorts/:cohortId';
 export const COHORT_PROJECTS_PAGE = '/cohorts/:cohortId/projects';
 export const COHORT_ALUMNI_PAGE = '/cohorts/:cohortId/alumni';
+export const SPECIFIC_ALUMNUS_PAGE = '/cohorts/:cohortId/alumni/:alumniId';
 export const SPECIFIC_PROJECT_PAGE = '/projects/:projectId';
 export const SPECIFIC_COHORT_PROJECT_PAGE =
   '/cohorts/:cohortId/projects/:projectId';
